requirements.txt
numba>=0.57.0
llvmlite>=0.40.0
umap-learn>=0.5.4
numpy<2.0.0
fastapi
uvicorn
celery
redis
requests
python-dotenv
edgartools
bertopic
hdbscan
